Completed issue 17 - show name
A name will now be shown in each normal display which will show the name of the
currently visible map (illustration) dependant on the showName and dmShowName
flags.  if the name is shown in DM mode, double clicking on it will toggle its
visiblity for the PC mode (i.e. shownName in database)